requirements.txt
certifi==2019.9.11
chardet==3.0.4
idna==2.8
numpy==1.17.3
pandas==0.25.2
plaid-python==3.4.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-http-client==3.2.1
pytz==2019.3
requests==2.22.0
sendgrid==6.1.0
six==1.12.0
urllib3==1.25.6
